% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_object_class}
\alias{check_object_class}
\title{Check if object is of specified class}
\usage{
check_object_class(object, class, accept_empty = TRUE)
}
\arguments{
\item{object}{Object to check}

\item{class}{Character vector of one or more expected classes}

\item{accept_empty}{Logical; if TRUE, empty/NULL objects pass check}

\item{variable_name}{Character; name of variable for error message. Default extracts from input.}
}
\value{
TRUE if object inherits from specified class(es), otherwise throws error
}
\description{
Checks whether an object inherits from specified class(es) and gives informative error message if not.
}
