% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_prep_refactored.R, R/ff_run.R, R/ff_sync.R
\docType{data}
\name{config}
\alias{config}
\title{Prepare data for XGBoost training, validating and predicting}
\format{
An object of class \code{list} of length 8.

An object of class \code{list} of length 8.

An object of class \code{list} of length 8.
}
\usage{
config

config

config
}
\arguments{
\item{datafolder}{Path to the main data directory, which should contain the "input" and "groundtruth" subdirectories.
These subdirectories must include the respective degree folders.Default is the system variable ff_datafolder.}

\item{tiles}{Vector of tiles in the syntax of e.g., "10N_080W" for which the data is prepared. Optional if either shape or country is given.}

\item{groundtruth_pattern}{Pattern to identify ground truth files. Default is "groundtruth6m" (set in config.json, groundtruth of future six months in binary format).}

\item{dates}{vector of dates in the format "YYYY-MM-DD". Default is "2021-01-01".}

\item{inc_features}{Vector of features to exclusively include in the data preparation.}

\item{exc_features}{Vector of features to exclude from the data preparation.}

\item{validation_sample}{Float between 0 and 1 indicating how much of the training dataset should be used for validation. Default is 0. Advised not to set above 0.3.}

\item{sample_size}{Fraction size of the random sample. Should be > 0 and <= 1. Default is 0.3.}

\item{adddate}{Boolean indicating whether to add date-related features ("sinmonth", "month","monthssince2019"). Default is TRUE.}

\item{shrink}{Option to modify the input area when a country is selected. This parameter determines how the spatial extent of the data is adjusted based on the selected country.
Options are:
\describe{
  \item{"none"}{No modification to the input area. The entire extent is used as-is. (Default)}
  \item{"crop"}{Crops the input area to the boundaries of the selected country using the exact extent of the country's shape.}
  \item{"crop-deg"}{Similar to "crop", but the resulting extent is adjusted to the nearest whole degree (latitude/longitude). This ensures the extent aligns with whole degree boundaries.}
  \item{"extract"}{Extracts data only within the boundaries of the selected country. The data is limited to the exact extent of the country's shape, and the function `terra::extract` is used to retrieve the data within this area.}
}}

\item{window}{Set the extent on which to process. Default is NA to derive it from the data.}

\item{label_threshold}{Threshold for labeling the ground truth. Default is 1, meaning a pixel is labeled as deforested if at least one deforestation event has occurred.
If `label_threshold = NA`, the ground truth will represent the total number of deforestation events rather than a binary label.}

\item{addxy}{Boolean indicating whether to add x and y coordinates as features. Default is FALSE.}

\item{shape}{SpatVector object representing the area of interest. Either `shape` or `country` must be provided.}

\item{country}{ISO3 country code. Either `shape` or `country` must be provided.}

\item{prediction_dates}{Dates for prediction in "YYYY-MM-DD" format.}

\item{train_dates}{The dates for which you want to create the training data. dates should be a vector with the format YYYY-MM-DD with DD being 01}

\item{validation_dates}{The dates for which you want to create a distinct validation matrix (if any, select validation = T to make a subsample of the same dates as the training data).}

\item{save_path}{Path to save the trained model (with extension ".model"). Default is NULL.}

\item{save_path_predictions}{Path to save the predictions (with extension ".tif"). Default is NULL.}

\item{trained_model}{Pre-trained model object or path to saved model. If NULL, a new model will be trained. Default is NULL.}

\item{ff_prep_params}{List of parameters for data preprocessing. See `ff_prep` function for details.}

\item{ff_train_params}{List of parameters for model training. See `ff_train` function for details.}

\item{threshold}{Probability threshold for binary classififf_cation. Default is 0.5.}

\item{fltr_features}{Feature dataset used for pre-filtering for training. Default is initialforestcover. Can be more than one}

\item{fltr_condition}{The condition with value that is used to filter the training dataset based on mask features. Default is ">0". Can be more than one}

\item{accuracy_csv}{Path to save accuracy metrics in CSV format. Default is NA (no CSV output).}

\item{importance_csv}{Path to save feature importance metrics in CSV format. Default is NA (no CSV output).}

\item{autoscale_sample}{Logical; Whether to automatically scale the number of samples based on the size of the area and the length of the training period.}

\item{validation}{Logical; Whether to add a validation matrix based on the training data, which is set at 0.25 of the training matrix. Should not be set if validation_dates is not NULL.}

\item{ff_folder}{Character. Local folder to sync data to.}

\item{identifier}{Character or SpatVector. When a character it should be either a tile ID (e.g., "00N_000E") or a country ISO3 code.}

\item{download_model}{Logical. Whether to download the corresponding model. Only works when downloading for entire countries. Default is FALSE.}

\item{download_data}{Logical. Whether to download the preprocessed input data. Default is TRUE.}

\item{download_groundtruth}{Logical. Whether to download the groundtruth data as well. This should be turned off when you want to use your own data as groundtruth. Default is TRUE.}

\item{download_predictions}{Logical. Whether to download the prediction data.Only works when downloading for entire countries. Default is FALSE.}

\item{bucket}{Character. Name of the S3 bucket. Default is "forestforesight-public".}

\item{region}{Character. AWS region of the bucket. Default is "eu-west-1".}

\item{verbose}{Logical. Whether the function should be verbose.}

\item{sync_verbose}{Logical. Whether the syncing should also be verbose. Default}
}
\value{
A list containing:
  \item{data_matrix}{A list with features and labels for training}
  \item{validation_matrix}{A list with features and labels for validation (if validation_sample > 0)}
  \item{testindices}{Indices of the filtered samples}
  \item{groundtruthraster}{A SpatRaster of the ground truth}
  \item{features}{A vector of feature names}
  \item{hasgroundtruth}{A boolean stating that the groundtruthraster is actually the groundtruth and not just a template}

A SpatRaster object containing the predicted deforestation probabilities.If multiple prediction dates are given you receive a rasterstack with a raster per date

Invisible NULL. The function is called for its side effects.
}
\description{
This function prepares data for the ForestForesight training and predicting algorithm based on specified parameters.

This function trains an XGBoost model using historical data and then predicts deforestation
on raster data for a specified date and area.

This function synchronizes ForestForesight data from a public S3 bucket to a local folder.
It can sync data for a specific tile or an entire country, including input data, ground truth data,
and optionally model data.
}
\examples{
\dontrun{
prepared_data <- ff_prep(
  datafolder = "path/to/data",
  country = "BRA",
  dates = ForestForesight::daterange("2022-01-01", "2022-12-31"),
  fltr_features = "initialforestcover",
  fltr_condition = ">0"
)
}

\dontrun{
# Predict deforestation for a country
prediction <- ff_run(
  country = "BRA",
  prediction_date = "2024-01-01",
  ff_folder = "path/to/forestforesight/data",
  train_dates = ForestForesight::daterange("2022-01-01", "2023-12-31"),
  save_path = "path/to/save/model.model",
  accuracy_csv = "path/to/save/accuracy.csv"
)

# Plot the prediction
plot(prediction)
}

\dontrun{
ff_sync("path/to/local/folder", "00N_000E")
ff_sync("path/to/local/folder", "BRA", download_model = TRUE)
}

}
\references{
Zillah Calle (2023)
Jonas van Duijvenbode (2023)

Jonas van Duijvenbode (2023)
Zillah Calle (2023)
}
\seealso{
\code{\link{ff_train}} for training a model with the prepared data
\code{\link{ff_predict}} for making predictions using a trained model

\code{\link{ff_prep}} for data preparation
\code{\link{ff_train}} for model training
\code{\link{ff_predict}} for making predictions
\code{\link{ff_analyze}} for analyzing prediction results
}
\keyword{data-preparation}
\keyword{datasets}
\keyword{forestry}
\keyword{machine-learning}
\keyword{prediction}
\keyword{raster}
