% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_train.R
\name{convert_to_dmatrix}
\alias{convert_to_dmatrix}
\title{Convert input matrix to XGBoost DMatrix format}
\usage{
convert_to_dmatrix(matrix)
}
\arguments{
\item{matrix}{Input matrix or DMatrix object}
}
\value{
XGBoost DMatrix object
}
\description{
Convert input matrix to XGBoost DMatrix format
}
