% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_matrix.R
\name{convert_to_matrix}
\alias{convert_to_matrix}
\title{Convert a list of files to a matrix}
\usage{
convert_to_matrix(files, daterange)
}
\arguments{
\item{files}{A character vector of file paths.}

\item{daterange}{A vector specifying the date range.}
}
\value{
A list containing matrices, where each matrix represents a feature.
}
\description{
This function converts a list of files into a matrix, where each row represents
a unique feature and each column represents a spatial pointer.
}
