% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_train.R
\name{create_params}
\alias{create_params}
\title{Create parameter list for XGBoost training}
\usage{
create_params(...)
}
\arguments{
\item{...}{Named parameters for XGBoost}
}
\value{
List of parameters
}
\description{
Create parameter list for XGBoost training
}
