% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_train.R
\name{create_watchlist}
\alias{create_watchlist}
\title{Create watchlist for model training}
\usage{
create_watchlist(train_dataset, validation_matrix)
}
\arguments{
\item{train_dataset}{Training DMatrix}

\item{validation_matrix}{Validation matrix or NULL}
}
\value{
Named list of matrices for monitoring
}
\description{
Create watchlist for model training
}
