% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{daterange}
\alias{daterange}
\title{Generate a vector with the first of every month between two given dates}
\usage{
daterange(start_date, end_date)
}
\arguments{
\item{start_date}{A character string representing the start date in "YYYY-MM-DD" format.}

\item{end_date}{A character string representing the end date in "YYYY-MM-DD" format.}
}
\value{
A character vector with the first of every month between start_date and end_date.
}
\description{
Generate a vector with the first of every month between two given dates
}
\examples{
daterange("2022-01-15", "2022-04-15")
# Returns: [1] "2022-01-01" "2022-02-01" "2022-03-01" "2022-04-01"

}
