% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfw_download.R
\name{download_gfw_alerts}
\alias{download_gfw_alerts}
\title{Download GFW Alerts Data}
\usage{
download_gfw_alerts(
  output_dir,
  country = NA,
  separate_folders = FALSE,
  basename = NA
)
}
\arguments{
\item{output_dir}{The directory where the downloaded data will be saved.}

\item{country}{An optional parameter specifying the ISO3 country code to filter the data. Default is NA.}

\item{separate_folders}{Logical. If TRUE, data is saved in separate folders based on tile_id. Default is FALSE.}

\item{basename}{An optional parameter to specify a basename for the downloaded files.
Applicable only when separate_folders is TRUE. Default is NA.}
}
\value{
locations of the created output tif-files.
}
\description{
Downloads GFW (Global Forest Watch) Alerts data and saves it to the specified output directory.
}
\examples{
\dontrun{
# Download GFW Alerts for a specific country and save in separate folders
download_gfw_alerts("path/to/output_directory", country = "SUR", separate_folders = TRUE, basename = "my_data")
}

}
