% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_analyse.R
\name{ff_analyse}
\alias{ff_analyse}
\title{Calculate Scores}
\usage{
ff_analyse(
  predictions,
  groundtruth,
  forestmask = NULL,
  csvfile = NULL,
  append = T,
  analysis_polygons = NULL,
  return_polygons = T,
  remove_empty = T
)
}
\arguments{
\item{predictions}{A character vector or raster object representing the predictions.}

\item{groundtruth}{A character vector or raster object representing the ground truth.}

\item{forestmask}{An optional character vector or raster object representing the forest mask.}

\item{csvfile}{An optional CSV file to which the results will be written.}

\item{append}{Logical. If TRUE, results will be appended to the existing CSV file.}

\item{analysis_polygons}{Optional vector or character vector representing analysis polygons.}

\item{return_polygons}{Logical. If TRUE, the polygons with calculated scores will be returned.}

\item{remove_empty}{Logical. If TRUE, empty rows (with all scores being zero) will be removed from the output.}
}
\value{
A vector dataset containing calculated scores for each polygon.
}
\description{
This function calculates scores based on predictions, ground truth, and optional parameters.
}
\examples{
\dontrun{
calculate_scores(predictions, groundtruth, forestmask, csvfile, append, analysis_polygons, return_polygons, remove_empty)
}

}
