% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_analyze_amounts.R
\name{ff_analyze_amounts}
\alias{ff_analyze_amounts}
\title{Calculate Scores}
\usage{
ff_analyze_amounts(
  predictions,
  groundtruth,
  forestmask = NULL,
  csvfile = NULL,
  country = NULL,
  append = TRUE,
  analysis_polygons = NULL,
  return_polygons = TRUE,
  remove_empty = TRUE,
  date = NULL,
  tile = NULL,
  method = NA,
  verbose = FALSE
)
}
\arguments{
\item{predictions}{A character vector or raster object representing the predictions.}

\item{groundtruth}{A character vector or raster object representing the ground truth.}

\item{forestmask}{An optional character vector or raster object representing the forest mask.}

\item{csvfile}{An optional CSV file to which the results will be written.}

\item{country}{Character. If NULL all the overlapping polygons will be processed.
Otherwise the ISO3 code should be given and the analysis_polygons dataset should contain a column called iso3}

\item{append}{Logical. If TRUE, results will be appended to the existing CSV file.}

\item{analysis_polygons}{Optional vector or character vector representing analysis polygons.}

\item{return_polygons}{Logical. If TRUE, the polygons with calculated scores will be returned.}

\item{remove_empty}{Logical. If TRUE, empty rows (with all scores being zero) will be removed from the output.}

\item{date}{character. should be in format (YYYY-MM-DD).
Optional if either groundtruth or predictions is a character to the tiffile.}

\item{tile}{character. should be in format AA{N-S}_BBB{W-E}.
Optional if either groundtruth or predictions is a character to the tiffile with a directory name.}

\item{method}{character. the shorthand for the method used,
which should also be included in the separate csv file for storing methods}

\item{verbose}{Logical. Whether the steps taken in the function should be verbose.}
}
\value{
A vector dataset containing calculated scores for each polygon.
}
\description{
This function calculates rmse scores based on predictions, ground truth, and optional parameters.
}
