% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_contextualize.R
\name{ff_contextualize}
\alias{ff_contextualize}
\title{contextualize hotzones or centroids}
\usage{
ff_contextualize(
  hotzones = NA,
  hotzonesfolder = NA,
  contextfolder,
  NA_label = "Unknown",
  country = NA,
  date = NA,
  outputfile = NA,
  return_vector = F,
  overwrite = T,
  WDPA = TRUE,
  GADM = TRUE,
  ECOBIOME = TRUE
)
}
\arguments{
\item{hotzones}{A character vector or raster object representing the predictions. Can be empty if country and date and datafolder are not empty}

\item{hotzonesfolder}{A path to the folder containing the tiles with predictions.}

\item{contextfolder}{A path to the folder that contains all the files that should be added as context features to the predictions}

\item{NA_label}{When the hotzone does not overlap with the context polygons, this is the value that should be used instead.}

\item{country}{A character vector of the iso3 code of the country that you want to process}

\item{date}{character. should be in format (YYYY-MM-DD). Optional if either groundtruth or predictions is a character to the tiffile.}

\item{outputfile}{character. The filename of the hotzone outputs}

\item{return_vector}{Logical. If TRUE, the SpatVector that was used as input is returned.}

\item{WDPA}{Logical. Whether the WDPA classification should be added. Default is TRUE}

\item{GADM}{Logical. Whether the GADM classification should be added. Default is TRUE}

\item{ECOBIOME}{Logical. Whether the ECOBIOME classifications should be added. Default is TRUE}

\item{overwrite.}{Whether existing polygons should be overwritten. default is TRUE}
}
\value{
A vector dataset containing the input data with all columns of the files in the context folder added
}
\description{
This function gives context for searchability in e.g. ArcGIS online or powerBI
}
