% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_dqc.R
\name{ff_dqc}
\alias{ff_dqc}
\title{Data quality control}
\usage{
ff_dqc(folder_path)
}
\arguments{
\item{folder_path}{The path to the folder containing TIF files.}
}
\description{
This function loops through all tif files in a given folder and gives a summary and total overview of the quality of the rasters
}
\examples{
# Example usage:
result <- process_tif_files("path/to/tif_folder")

}
