% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_dqcfile.R
\name{ff_dqc_file}
\alias{ff_dqc_file}
\title{summary data aspects of a raster}
\usage{
ff_dqc_file(raster, return_values = T)
}
\arguments{
\item{raster}{path to raster.}

\item{return_values}{Should the values of the rasters also be returned.}
}
\description{
requires a path to a raster as input and gives some quality parameters of that file
}
\examples{
# Example usage:
result <- process_tif_files("path/to/tif_file")

}
