% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_predict.R
\name{ff_predict}
\alias{ff_predict}
\title{Predict with a trained XGBoost model}
\usage{
ff_predict(
  model,
  test_matrix,
  threshold = 0.5,
  groundtruth = NA,
  indices = NA,
  templateraster = NA,
  verbose = T,
  certainty = F
)
}
\arguments{
\item{model}{The trained XGboost model}

\item{test_matrix}{The xgbd test matrix}

\item{threshold}{Vector with chosen threshold(s). Default 0.5, which has shown to be the best in almost all scenarios}

\item{groundtruth}{A vector of the same length as the test matrix to verify against}

\item{indices.}{a vector of the indices of the template raster that need to be filled in.}

\item{templateraster.}{a spatraster that can serve as the template to fill in the predictions}

\item{verbose.}{whether the output should be verbose}

\item{certainty.}{if True the certainty in percentage of the prediction will be returned, otherwise just true or false}
}
\value{
list with precision, recall and F0.5
}
\description{
This function predicts on a trained xgboost model and a test dataset. it returns the precision, recall and F0.5
}
\references{
Jonas van Duijvenbode (2023)
}
\keyword{XGBoost}
\keyword{data}
\keyword{preparation}
