% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_prep.R
\name{ff_prep}
\alias{ff_prep}
\title{Prepare data for XGBoost training, validating and predicting}
\usage{
ff_prep(
  datafolder = NA,
  country = NA,
  shape = NA,
  tiles = NULL,
  groundtruth_pattern = "groundtruth6m",
  start = "2023-01-01",
  end = NA,
  inc_features = NA,
  exc_features = NA,
  fltr_features = NULL,
  fltr_condition = NULL,
  sample_size = 0.3,
  validation_sample = 0,
  adddate = T,
  sampleraster = T,
  verbose = T,
  shrink = "none",
  window = NA,
  label_threshold = 1,
  addxy = F
)
}
\arguments{
\item{datafolder}{Path to the data folder. Should contain the degrees folders. Default is the system variable ff_datafolder.}

\item{country}{ISO3 code of the country or countries for which the data is prepared. Optional if either shape or tiles is given.}

\item{shape}{SpatVector for which the data is prepared. Optional if either country or tiles is given.}

\item{tiles}{Vector of tiles in the syntax of e.g., "10N_080W". Optional if either shape or country is given.}

\item{groundtruth_pattern}{Pattern to identify ground truth files. Default is "groundtruth6m" (groundtruth of future six months in binary format).}

\item{start}{Start date for training data in the format "YYYY-MM-DD". Default is "2021-01-01".}

\item{end}{End date for training data in the format "YYYY-MM-DD". Default is NA to only process the start month.}

\item{inc_features}{Vector of features to include in the data preparation.}

\item{exc_features}{Vector of features to exclude from the data preparation.}

\item{fltr_features}{Vector of features for filtering data. Default is NULL. Example: 'initialforestcover'.}

\item{fltr_condition}{Vector of filtering conditions. Default is NULL. Example: '>0'. Should consist of operator and value.}

\item{sample_size}{Fraction size of the random sample. Should be > 0 and <= 1. Default is 0.3.}

\item{validation_sample}{Float between 0 and 1 indicating how much of the training dataset should be used for validation. Default is 0. Advised not to set above 0.3.}

\item{adddate}{Boolean indicating whether to add date-related features. Default is TRUE.}

\item{sampleraster}{Boolean indicating if sampling raster should be used. Default is TRUE.}

\item{verbose}{Boolean indicating whether to display progress messages. Default is TRUE.}

\item{shrink}{Option to shrink the input area if a country was selected. Options: "none", "crop", "crop-deg", "extract". Default is "none".}

\item{window}{Set the extent on which to process. Default is NA to derive it from the data.}

\item{label_threshold}{Threshold for labeling. Default is 1.}

\item{addxy}{Boolean indicating whether to add x and y coordinates as features. Default is FALSE.}
}
\value{
A list containing:
  \item{data_matrix}{A list with features and labels for training}
  \item{validation_matrix}{A list with features and labels for validation (if validation_sample > 0)}
  \item{testindices}{Indices of the filtered samples}
  \item{groundtruthraster}{A SpatRaster of the ground truth}
  \item{features}{A vector of feature names}
}
\description{
This function prepares data for the ForestForesight training and predicting algorithm based on specified parameters.
}
\examples{
\dontrun{
prepared_data <- ff_prep(
  datafolder = "path/to/data",
  country = "BRA",
  start = "2022-01-01",
  end = "2022-12-31",
  fltr_features = "initialforestcover",
  fltr_condition = ">0"
)
}

}
\references{
Jonas van Duijvenbode (2023)
Zillah Calle (2023)
}
\seealso{
\code{\link{ff_train}} for training a model with the prepared data
\code{\link{ff_predict}} for making predictions using a trained model
}
\keyword{data-preparation}
\keyword{forestry}
\keyword{machine-learning}
