% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_structurecheck.R
\name{ff_structurecheck}
\alias{ff_structurecheck}
\title{Check and Report ForestForesight File Structure}
\usage{
ff_structurecheck(shape, folder_path, check_date = NULL)
}
\arguments{
\item{shape}{A SpatVector object representing the area of interest.}

\item{folder_path}{Character string. Path to the main ForestForesight folder.}

\item{check_date}{Character string. Date to check for in the format "YYYY-MM-DD".
If NULL, uses the first of the current month.}
}
\value{
Invisible NULL. The function is called for its side effects (printing).
}
\description{
This function checks the file structure for ForestForesight data based on a given
spatial vector and folder path. It reports on the presence and correctness of
various folders and files required for ForestForesight operations.
}
