% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_sync.R
\name{ff_sync}
\alias{ff_sync}
\title{Sync ForestForesight Data from S3}
\usage{
ff_sync(
  ff_folder,
  identifier,
  download_model = FALSE,
  download_data = TRUE,
  download_predictions = FALSE,
  download_groundtruth = TRUE,
  bucket = "forestforesight-public",
  region = "eu-west-1",
  verbose = TRUE,
  sync_verbose = FALSE
)
}
\arguments{
\item{ff_folder}{Character. Local folder to sync data to.}

\item{identifier}{Character or SpatVector. When a character it should be either a tile ID (e.g., "00N_000E") or a country ISO3 code.}

\item{download_model}{Logical. Whether to download the corresponding model. Only works when downloading for entire countries. Default is FALSE.}

\item{download_data}{Logical. Whether to download the preprocessed input data. Default is TRUE.}

\item{download_predictions}{Logical. Whether to download the prediction data.Only works when downloading for entire countries. Default is FALSE.}

\item{download_groundtruth}{Logical. Whether to download the groundtruth data as well. This should be turned off when you want to use your own data as groundtruth. Default is TRUE.}

\item{bucket}{Character. Name of the S3 bucket. Default is "forestforesight-public".}

\item{region}{Character. AWS region of the bucket. Default is "eu-west-1".}

\item{verbose}{Logical. Whether the function should be verbose.}

\item{sync_verbose}{Logical. Whether the syncing should also be verbose. Default}
}
\value{
Invisible NULL. The function is called for its side effects.
}
\description{
This function synchronizes ForestForesight data from a public S3 bucket to a local folder.
It can sync data for a specific tile or an entire country, including input data, ground truth data,
and optionally model data.
}
\examples{
\dontrun{
ff_sync("path/to/local/folder", "00N_000E")
ff_sync("path/to/local/folder", "BRA", download_model = TRUE)
}

}
