% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_train.R
\name{ff_train}
\alias{ff_train}
\title{Train an XGBoost Model for ForestForesight}
\usage{
ff_train(
  train_matrix,
  validation_matrix = NULL,
  nrounds = 200,
  eta = 0.1,
  max_depth = 5,
  subsample = 0.75,
  eval_metric = "aucpr",
  early_stopping_rounds = 10,
  gamma = NULL,
  maximize = NULL,
  min_child_weight = 1,
  verbose = FALSE,
  xgb_model = NULL,
  modelfilename = NULL,
  objective = "binary:logistic"
)
}
\value{
A trained XGBoost model (xgb.Booster object).
}
\description{
This function trains an XGBoost model with optimized default
parameters derived from worldwide data analysis.
}
