% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_feature.R
\name{filter_by_feature}
\alias{filter_by_feature}
\title{Filter matrix by feature}
\usage{
filter_by_feature(fltr_features, fltr_condition, matrix, verbose = T)
}
\arguments{
\item{fltr_features}{A character vector specifying the features to filter.}

\item{fltr_condition}{A character vector specifying the conditions for each feature.}

\item{matrix}{The input matrix to filter.}

\item{verbose}{Logical indicating whether to display verbose output (default is TRUE).}
}
\value{
A list containing the filtered matrix and the indices of the filtered rows.
}
\description{
This function filters a matrix based on specified features and conditions.
}
\examples{
filter_by_feature(
  fltr_features = c("landpercentage", "forestmask"),
  fltr_condition = c(">100", ">0"),
  matrix = my_matrix
)

}
