% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_threshold.R
\name{find_best_threshold}
\alias{find_best_threshold}
\title{Find Best Threshold}
\usage{
find_best_threshold(
  prediction,
  groundtruth,
  optimize_function = get_f_score,
  a = 0.45,
  b = 0.55,
  tol = 0.001,
  maxiter = 100,
  beta = 0.5
)
}
\arguments{
\item{prediction}{A vector of predictions (numeric) or SpatRaster.}

\item{groundtruth}{A vector of ground truth values (binary) or SpatRaster.}

\item{optimize_function}{The evaluation function to optimize. Default is get_f_score.}

\item{a}{Initial guess for the lower bound of threshold search.}

\item{b}{Initial guess for the upper bound of threshold search.}

\item{tol}{Tolerance for convergence.}

\item{maxiter}{Maximum number of iterations.}

\item{beta}{The weight of precision in the F-score calculation.}
}
\value{
A list containing the best threshold and the corresponding F-score.
}
\description{
This function finds the best threshold for a given prediction function by maximizing the evaluation function.
}
\examples{
find_best_threshold(c(0.2, 0.6, 0.7), c(0, 1, 1))
}
