% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Get Information About a Spatial Shape or Country}
\usage{
get_info(shape_or_iso, ff_dir = NULL, verbose = TRUE)
}
\arguments{
\item{shape_or_iso}{SpatVector object or 3-digit ISO code}

\item{ff_dir}{Path to ForestForesight data directory. Default: NULL}

\item{verbose}{If TRUE, prints information. Default: TRUE}
}
\value{
List containing spatial information (invisibly)
}
\description{
Get Information About a Spatial Shape or Country
}
