% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasvalue.R
\name{hasvalue}
\alias{hasvalue}
\title{Check if an object has a value and is not NULL or NA}
\usage{
hasvalue(x)
}
\arguments{
\item{x}{An R object to be tested.}
}
\value{
TRUE if the object is not NULL or NA, else FALSE. returns TRUE if it is a vector of multiple NA's
}
\description{
This function checks whether an object is not NULL or NA.
}
\examples{
# Test with a vector containing NA values
hasvalue(c(1, 2, NA, 4))
# Test with a NULL object
hasvalue(NULL)

}
