% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_lists.R
\name{merge_lists}
\alias{merge_lists}
\title{Merge Two Named Lists with User Preferences Overwriting Defaults}
\usage{
merge_lists(default, user)
}
\arguments{
\item{default}{A named list containing the default values.}

\item{user}{A named list containing the user-specified values that may overwrite the default values.}
}
\value{
A named list with combined elements from both input lists. User-specified elements
overwrite default elements with the same names.
}
\description{
This function merges two named lists, allowing elements from the user-specified list
to overwrite those in the default list if they have the same names. If there are no
overlapping keys, the function simply combines the lists.
}
\examples{
default_list <- list(a = 1, b = 2, c = 3)
user_list <- list(b = 20, d = 4)
merged_list <- merge_lists(default_list, user_list)
print(merged_list) # Should print: $a [1], $c [3], $b [20], $d [4]

}
