% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_accuracyreport.R
\name{prepare_importance_data}
\alias{prepare_importance_data}
\title{Prepare Importance Data for Plotting}
\usage{
prepare_importance_data(importance_results)
}
\arguments{
\item{importance_results}{Data frame of importance results}
}
\value{
Data frame of processed importance data
}
\description{
Prepare Importance Data for Plotting
}
\keyword{internal}
