% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_train.R
\name{save_model}
\alias{save_model}
\title{Save trained model to file}
\usage{
save_model(model, filename, verbose = FALSE)
}
\arguments{
\item{model}{Trained XGBoost model}

\item{filename}{Output filename}

\item{verbose}{Whether to print progress messages}
}
\description{
Save trained model to file
}
