% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_files_date}
\alias{select_files_date}
\title{Select Files Based on Closest Previous Date for Each Feature}
\usage{
select_files_date(given_date, listed_files)
}
\arguments{
\item{given_date}{The reference date in the format "YYYY-MM-DD".}

\item{listed_files}{A character vector of file names.}
}
\value{
A character vector of selected file names based on the closest previous date for each feature.
}
\description{
This function takes a given date and a vector of file names and returns a vector
containing the file names corresponding to the closest previous date for each unique feature.
}
\examples{
\dontrun{
# Example usage:
given_date <- "2022-03-15"
listed_files <- c(
  "mydir/tile_YYYY-MM-DD_feature.tif",
  "mydir/tile_YYYY-MM-DD_feature.tif",
  "mydir/tile_YYYY-MM-DD_feature.tif",
  "mydir/tile_YYYY-MM-DD_feature.tif",
  "mydir/tile_YYYY-MM-DD_feature.tif",
)

result <- select_files_date(given_date, listed_files)
print(result)
}

}
\seealso{
\code{\link{grep}}, \code{\link{lubridate::ymd}}
}
