% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_check_file.R
\name{summarize_rast}
\alias{summarize_rast}
\title{summary data aspects of a raster}
\usage{
summarize_rast(raster)
}
\arguments{
\item{raster}{path to raster.}
}
\description{
requires a path to a raster as input and gives some quality parameters of that file
}
\examples{
# Example usage:
result <- process_tif_files("path/to/tif_file")

}
