% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_feature_model_match.R
\name{test_feature_model_match}
\alias{test_feature_model_match}
\title{Test Feature and Model Match}
\usage{
test_feature_model_match(model, feature_names = NULL)
}
\arguments{
\item{model}{Either a character string representing the path to the XGBoost model file, or an `xgb.Booster` object.}

\item{feature_names}{A character vector of feature names. This parameter should be provided if `model` is an `xgb.Booster` object. If `model` is a file path, the function will look for a corresponding `.rda` file with feature names in the same directory.}
}
\value{
A logical value: `TRUE` if the feature names match the features in the model, `FALSE` otherwise.
}
\description{
This function tests whether the provided feature names match the features in the given XGBoost model. It checks if the model file and the corresponding feature names file exist, and if so, loads them. It then attempts to generate the importance matrix to validate the match.
}
\examples{
\dontrun{
# Example with model file and feature names file
model_file <- "path/to/model.model"
feature_names_file <- "path/to/model.rda"
load(feature_names_file)
test_feature_model_match(model_file)

# Example with an xgb.Booster object
model <- xgboost::xgb.train(params = list(objective = "binary:logistic"), data = dtrain, nrounds = 10)
feature_names <- colnames(dtrain)
test_feature_model_match(model, feature_names)
}

}
