% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_train.R
\name{train_model}
\alias{train_model}
\title{Train XGBoost model}
\usage{
train_model(params, train_dataset, xgb_model = xgb_model, ...)
}
\arguments{
\item{params}{Model parameters}

\item{...}{Additional parameters passed to xgb.train}

\item{dtrain}{Training data}
}
\value{
Trained XGBoost model
}
\description{
Train XGBoost model
}
