% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_train.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate input data}
\usage{
validate_inputs(train_matrix)
}
\arguments{
\item{train_matrix}{Training data}
}
\value{
NULL invisibly
}
\description{
Validate input data
}
