% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_cat.R
\name{ff_cat}
\alias{ff_cat}
\title{Colorized Cat Function}
\usage{
ff_cat(
  ...,
  color = NULL,
  sep = " ",
  fill = FALSE,
  labels = NULL,
  append = FALSE,
  style = NULL,
  verbose = FALSE,
  logfile = NULL,
  timestamp = FALSE,
  auto_newline = TRUE
)
}
\arguments{
\item{...}{Arguments to be concatenated and printed.}

\item{color}{Character string specifying the color to apply to the text.
Must be a valid crayon color (e.g., "red", "blue", "green").}

\item{sep}{Character string to separate the arguments when concatenating.}

\item{fill}{Logical. If TRUE, a newline is appended after the output.}

\item{labels}{Character vector of labels for the lines printed.}

\item{append}{Logical. If TRUE, output will be appended to the file.}

\item{style}{Character string specifying additional style to apply.
Must be a valid crayon style (e.g., "bold", "underline").}

\item{logfile}{Character. Path to a logfile that should be printed to.}

\item{timestamp}{Logical. Whether a timestamp should be added}

\item{auto_newline}{Logical. Whether a newline should be added}

\item{verbose.}{Logical. Whether the ff_cat function should print. Is always overridden by logfile if that is set}
}
\value{
No return value, called for side effects.
}
\description{
This function wraps the base R `cat()` function and adds color and style
capabilities using the crayon package.
}
\examples{
ff_cat("Hello, world!")
ff_cat("This is red text", color = "red")
ff_cat("The sky is", "blue", color = "blue")
ff_cat("Bold green text", color = "green", style = "bold")
ff_cat("Line 1", "Line 2", color = "cyan", fill = TRUE)

}
