% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_polygonize.R
\name{ff_polygonize}
\alias{ff_polygonize}
\title{Polygonize raster to risk areas}
\usage{
ff_polygonize(
  raster,
  output_file = NA,
  min_pixel = 5,
  threshold = 0.5,
  window_size = 7,
  smoothness = 2,
  verbose = FALSE,
  calc_max = FALSE,
  contain_polygons = NA
)
}
\arguments{
\item{raster}{Character or SpatRaster. Path to the input raster file or a SpatRaster.}

\item{output_file}{Character. File path to save shapefiles.
Default is NA and means the polygons are returned and not written.}

\item{min_pixel}{Numeric. Base pixel count to filter on. Default is 5}

\item{threshold}{Numeric or Character. Threshold to apply to the raster. Default is 0.5.
Options are:
\describe{
  \item{"numeric"}{A value between 0 and 1 that reflects the value that you want as a threshold}
  \item{"medium"}{Automatically chosen threshold for medium risk}
  \item{"high"}{Automatically chosen threshold for medium risk}
  \item{"very high"}{Automatically chosen threshold for medium risk}
}}

\item{window_size}{Numeric. Window size for focal calculation. Default is 7.}

\item{smoothness}{Numeric. Smoothness parameter for ksmooth method. Default is 2}

\item{verbose}{Logical. Whether the automatically detected threshold should be plotted}

\item{calc_max}{Logical.
If this is set to True a reasonable amount of polygons will be outputted depending on the area size}

\item{contain_polygons}{Logical.
If calc_max is enabled, only polygons will be created that lie within the SpatVector contain_polygons}
}
\value{
A SpatVector object containing the processed and simplified polygons.
}
\description{
This function processes a raster file, applies thresholds, creates polygons,
and simplifies them based on various parameters.
}
\examples{
\dontrun{
ff_polygonize("path/to/raster.tif", "output_polygons.shp",
  min_pixel = 5, threshold = 0.5, window_size = 7
)
}

}
\references{
Jonas van Duijvenbode (2023)
Zillah Calle (2023)
}
