% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_prep.R
\name{ff_prep}
\alias{ff_prep}
\title{Prepare data for XGBoost training, validating and predicting}
\usage{
ff_prep(
  datafolder = Sys.getenv("DATA_FOLDER"),
  country = NA,
  shape = NA,
  tiles = NULL,
  groundtruth_pattern = Sys.getenv("DEFAULT_GROUNDTRUTH"),
  dates = "2023-01-01",
  inc_features = NA,
  exc_features = NA,
  filter_features = NULL,
  filter_conditions = NULL,
  sample_size = 0.3,
  validation_sample = 0,
  add_date = TRUE,
  verbose = TRUE,
  shrink = "none",
  window = NA,
  label_threshold = 1,
  add_xy = FALSE
)
}
\arguments{
\item{datafolder}{Path to the main data directory,
which should contain the "input" and "groundtruth" subdirectories.
These subdirectories must include the respective degree folders.
Default is the system variable ff_datafolder.}

\item{country}{ISO3 code of the country or countries for which the data is prepared.
Optional if either shape or tiles is given.}

\item{shape}{SpatVector for which the data is prepared.
Optional if either country or tiles is given.}

\item{tiles}{Vector of tiles in the syntax of e.g., "10N_080W" for which the data is prepared.
Optional if either shape or country is given.}

\item{groundtruth_pattern}{Pattern to identify ground truth files.
Default is "groundtruth6m" (set in config.json, groundtruth of future six months in binary format).}

\item{dates}{vector of dates in the format "YYYY-MM-DD". Default is "2021-01-01".}

\item{inc_features}{Vector of features to exclusively include in the data preparation.}

\item{exc_features}{Vector of features to exclude from the data preparation.}

\item{filter_features}{Vector of features for filtering data. Default is NULL.
Example: 'initialforestcover'.}

\item{filter_conditions}{Vector of conditions corresponding to `filter_features`.
Each entry should consist of an operator and a value.
Should be of the same length as `filter_features`. Default is NULL. Example: '>0'.}

\item{sample_size}{Fraction size of the random sample. Should be > 0 and <= 1. Default is 0.3.}

\item{validation_sample}{Float between 0 and 1 indicating how much of the
training dataset should be used for validation. Default is 0.
Advised not to set above 0.3.}

\item{add_date}{Boolean indicating whether to add date-related features
("sinmonth", "month","monthssince2019"). Default is TRUE.}

\item{verbose}{Boolean indicating whether to display progress messages. Default is TRUE.}

\item{shrink}{Option to modify the input area when a country is selected.
This parameter determines how the spatial extent of the data is
adjusted based on the selected country.
Options are:
\describe{
  \item{"none"}{No modification to the input area. The entire extent is used as-is. (Default)}
  \item{"crop"}{Crops the input area to the boundaries of the selected country
  using the exact extent of the country's shape.}
  \item{"crop-deg"}{Similar to "crop", but the resulting extent is adjusted
  to the nearest whole degree (latitude/longitude). This ensures the extent aligns with whole degree boundaries.}
  \item{"extract"}{Extracts data only within the boundaries of the selected country.
  The data is limited to the exact extent of the country's shape,
  and the function `terra::extract` is used to retrieve the data within this area.}
}}

\item{window}{Set the extent on which to process. Default is NA to derive it from the data.}

\item{label_threshold}{Threshold for labeling the ground truth.
Default is 1, meaning a pixel is labeled as deforested if at least one deforestation event has occurred.
If `label_threshold = NA`, the ground truth will represent the total number of
deforestation events rather than a binary label.}

\item{add_xy}{Boolean indicating whether to add x and y coordinates as features. Default is FALSE.}
}
\value{
A list containing:
  \item{feature_dataset}{A list with features and labels for training}
  \item{validation_matrix}{A list with features and labels for validation (if validation_sample > 0)}
  \item{test_indices}{Indices of the filtered samples}
  \item{groundtruth_raster}{A SpatRaster of the ground truth}
  \item{features}{A vector of feature names}
  \item{has_groundtruth}{A boolean stating that the groundtruthraster is actually the groundtruth and not just a template}
}
\description{
This function prepares data for the ForestForesight training and predicting algorithm based on specified parameters.
}
\examples{
\dontrun{
prepared_data <- ff_prep(
  datafolder = "path/to/data",
  country = "BRA",
  dates = ForestForesight::daterange("2022-01-01", "2022-12-31"),
  filter_features = "initialforestcover",
  filter_conditions = ">0"
)
}

}
\references{
Zillah Calle (2023)
Jonas van Duijvenbode (2023)
}
\seealso{
\code{\link{ff_train}} for training a model with the prepared data
\code{\link{ff_predict}} for making predictions using a trained model
}
\keyword{data-preparation}
\keyword{forestry}
\keyword{machine-learning}
