% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_run.R
\name{ff_run}
\alias{ff_run}
\title{Train a Model and Predict Deforestation on Raster Data}
\usage{
ff_run(
  shape = NULL,
  country = get_variable("DEFAULT_COUNTRY"),
  prediction_dates = NULL,
  ff_folder = get_variable("FF_FOLDER"),
  train_dates = NULL,
  validation_dates = NULL,
  model_save_path = NULL,
  predictions_save_path = NULL,
  pretrained_model_path = NULL,
  ff_prep_parameters = NULL,
  ff_train_parameters = NULL,
  certainty_threshold = get_variable("DEFAULT_THRESHOLD"),
  filter_features = get_variable("FOREST_MASK"),
  filter_conditions = get_variable("FOREST_MASK_FILTER"),
  accuracy_output_path = NULL,
  importance_output_path = NULL,
  verbose = TRUE,
  autoscale_sample = FALSE,
  validation = FALSE,
  groundtruth_pattern = get_variable("DEFAULT_GROUNDTRUTH")
)
}
\arguments{
\item{shape}{SpatVector object representing the area of interest. Either `shape` or `country` must be provided.}

\item{country}{ISO3 country code. Either `shape` or `country` must be provided.}

\item{prediction_dates}{Dates for prediction in "YYYY-MM-DD" format.}

\item{ff_folder}{Directory containing the input data.}

\item{train_dates}{The dates for which you want to create the training data.
dates should be a vector with the format YYYY-MM-DD with DD being 01}

\item{validation_dates}{The dates for which you want to create a distinct validation matrix
(if any, select validation = T to make a subsample of the same dates as the training data).}

\item{model_save_path}{Path to save the trained model (with extension ".model"). Default is NULL.}

\item{predictions_save_path}{Path to save the predictions (with extension ".tif"). Default is NULL.}

\item{pretrained_model_path}{Pre-trained model object or path to saved model.
If NULL, a new model will be trained. Default is NULL.}

\item{ff_prep_parameters}{List of parameters for data preprocessing. See `ff_prep` function for details.}

\item{ff_train_parameters}{List of parameters for model training. See `ff_train` function for details.}

\item{certainty_threshold}{Probability certainty_threshold for binary classification. Default is 0.5.}

\item{filter_features}{Feature dataset used for pre-filtering for training.
Default is initialforestcover. Can be more than one}

\item{filter_conditions}{The condition with value that is used to filter the training dataset based on mask features.
Default is ">0". Can be more than one}

\item{accuracy_output_path}{Path to save accuracy metrics in CSV format. Default is NULL (no CSV output).}

\item{importance_output_path}{Path to save feature importance metrics in CSV format. Default is NULL (no CSV output).}

\item{verbose}{Logical; whether to display progress messages. Default is TRUE.}

\item{autoscale_sample}{Logical; Whether to automatically scale the number of samples
based on the size of the area and the length of the training period.}

\item{validation}{Logical; Whether to add a validation matrix based on the training data,
which is set at 0.25 of the training matrix. Should not be set if validation_dates is not NULL.}
}
\value{
A list containing:
  \item{prediction_timeseries}{A SpatRaster object or RasterStack containing the predicted deforestation
        probabilities for each prediction date}
  \item{shape}{The SpatVector object used for analysis, either from direct input or derived from country code}
  \item{model}{Path to the model used for predictions (either newly trained or pretrained)}
  \item{accuracy_dataframe}{A SpatialPolygonsDataFrame containing accuracy metrics for each analyzed area.
        NULL if no groundtruth data was available}
  \item{importance_dataframe}{A data frame containing feature importance metrics.
        NULL if importance calculation was not requested}
}
\description{
This function trains an XGBoost model using historical data and then predicts deforestation
on raster data for a specified date and area.
}
\examples{
\dontrun{
# Predict deforestation for a country
prediction <- ff_run(
  country = "BRA",
  prediction_date = "2024-01-01",
  ff_folder = "path/to/forestforesight/data",
  train_dates = ForestForesight::daterange("2022-01-01", "2023-12-31"),
  model_save_path = "path/to/save/model.model",
  accuracy_output_path = "path/to/save/accuracy.csv"
)

# Plot the prediction
plot(prediction)
}

}
\references{
Jonas van Duijvenbode (2023)
Zillah Calle (2023)
}
\seealso{
\code{\link{ff_prep}} for data preparation
\code{\link{ff_train}} for model training
\code{\link{ff_predict}} for making predictions
\code{\link{ff_analyze}} for analyzing prediction results
}
\keyword{forestry}
\keyword{machine-learning}
\keyword{prediction}
\keyword{raster}
