% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_run.R
\name{ff_run}
\alias{ff_run}
\title{Train a Model and Predict Deforestation on Raster Data}
\usage{
ff_run(
  shape = NULL,
  country = NULL,
  prediction_dates,
  ff_folder,
  train_start = NULL,
  train_end = NULL,
  save_path = NULL,
  trained_model = NULL,
  ff_prep_params = NULL,
  ff_train_params = NULL,
  threshold = 0.5,
  mask_feature = "initialforestcover",
  fltr_condition = ">0",
  accuracy_csv = NA,
  overwrite = F,
  verbose = T
)
}
\arguments{
\item{shape}{SpatVector object representing the area of interest. Either `shape` or `country` must be provided.}

\item{country}{ISO3 country code. Either `shape` or `country` must be provided.}

\item{prediction_dates}{Dates for prediction in "YYYY-MM-DD" format.}

\item{ff_folder}{Directory containing the input data.}

\item{train_start}{Start date for training data in "YYYY-MM-DD" format. Default is NULL.}

\item{train_end}{End date for training data in "YYYY-MM-DD" format. Default is NULL.}

\item{save_path}{Path to save the trained model. Default is NULL.}

\item{trained_model}{Pre-trained model object or path to saved model. If NULL, a new model will be trained. Default is NULL.}

\item{ff_prep_params}{List of parameters for data preprocessing. See `ff_prep` function for details.}

\item{ff_train_params}{List of parameters for model training. See `ff_train` function for details.}

\item{threshold}{Probability threshold for binary classification. Default is 0.5.}

\item{mask_feature}{Feature dataset used for pre-filtering for training. Default is initialforestcover. Can be more than one}

\item{fltr_condition}{The condition with value that is used to filter the training dataset based on mask features. Default is ">0". Can be more than one}

\item{accuracy_csv}{Path to save accuracy metrics in CSV format. Default is NA (no CSV output).}

\item{overwrite}{Logical; whether to overwrite existing files. Default is FALSE.}

\item{verbose}{Logical; whether to display progress messages. Default is TRUE.}
}
\value{
A SpatRaster object containing the predicted deforestation probabilities.If multiple prediction dates are given you receive a rasterstack with a raster per date
}
\description{
This function trains an XGBoost model using historical data and then predicts deforestation
on raster data for a specified date and area.
}
\examples{
\dontrun{
# Predict deforestation for a country
prediction <- ff_run(
  country = "BRA",
  prediction_date = "2024-01-01",
  ff_folder = "path/to/forestforesight/data",
  train_start = "2022-01-01",
  train_end = "2023-12-31",
  save_path = "path/to/save/model.model",
  accuracy_csv = "path/to/save/accuracy.csv"
)

# Plot the prediction
plot(prediction)
}

}
\references{
Jonas van Duijvenbode (2023)
Zillah Calle (2023)
}
\seealso{
\code{\link{ff_prep}} for data preparation
\code{\link{ff_train}} for model training
\code{\link{ff_predict}} for making predictions
\code{\link{ff_analyze}} for analyzing prediction results
}
\keyword{forestry}
\keyword{machine-learning}
\keyword{prediction}
\keyword{raster}
