% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_run.R
\name{ff_run}
\alias{ff_run}
\title{Run ForestForesight prediction on specified tiles}
\usage{
ff_run(
  datafolder,
  tiles,
  model = NULL,
  start = c(2021, 1),
  end = c(2021, 12),
  pstart = c(2022, 5),
  pend = c(2023, 5),
  savemodel = T,
  savefolder = NULL,
  analyse = T,
  overwrite = F,
  csvfile = NULL,
  append_to_csv = T
)
}
\arguments{
\item{datafolder}{Path to the folder containing the input data.}

\item{tiles}{Vector of tile names to process (e.g. 10N_080W).}

\item{model}{Trained model object. If NULL, a new model will be trained.}

\item{start}{Start date for training data (default: c(2021, 1)).}

\item{end}{End date for training data (default: c(2021, 12)).}

\item{pstart}{Start date for prediction (default: c(2022, 5)).}

\item{pend}{End date for prediction (default: c(2023, 5)).}

\item{savemodel}{Logical indicating whether to save the trained model (default: TRUE).}

\item{savefolder}{Path to the folder where models and predictions will be saved.}

\item{analyse}{Logical indicating whether to perform analysis on interest polygons (default: TRUE).}

\item{overwrite}{Logical indicating whether to overwrite existing prediction files (default: FALSE).}

\item{csvfile}{Path to the CSV file for storing analysis results.}

\item{append_to_csv}{Logical indicating whether to append results to the existing CSV file (default: TRUE for append, else overwrite).}
}
\value{
None
}
\description{
Run ForestForesight prediction on specified tiles
}
