% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_visualize.R
\name{ff_visualize}
\alias{ff_visualize}
\title{visualize hotzones}
\usage{
ff_visualize(
  predictions = NA,
  predfolder = NA,
  country = NA,
  date = NA,
  outputfile,
  t_cutoff = 0.5,
  return_polygons = F,
  overwrite = T
)
}
\arguments{
\item{predictions}{A character vector or raster object representing the predictions.can be empty if country and date and predfolder are not empty}

\item{predfolder}{A path to the folder containing the tiles with predictions.}

\item{country}{A character vector of the iso3 code of the country that you want to process}

\item{date}{character. should be in format (YYYY-MM-DD). Optional if either groundtruth or predictions is a character to the tiffile.}

\item{outputfile}{character. The filename of the hotzone outputs}

\item{t_cutoff}{The cutoff value below which no polygons should be returned. normally set at 0.5}

\item{return_polygons}{Logical. If TRUE, the hotzone polygons.}

\item{overwrite.}{Whether existing polygons should be overwritten. default is TRUE}
}
\value{
A vector dataset containing calculated scores for each polygon.
}
\description{
This function creates the hotzones of 400x400 meters with accuracy built in
}
