% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{This function provides information about a given spatial shape or country,
including the number of tiles it covers, available features, area, bounding box,
overlapping countries, and country groups.}
\usage{
get_info(shape_or_iso, ff_dir = NULL, verbose = TRUE)
}
\arguments{
\item{shape_or_iso}{SpatVector object or 3-digit ISO code}

\item{ff_dir}{Path to ForestForesight data directory. Default: NULL}

\item{verbose}{If TRUE, prints information. Default: TRUE}
}
\value{
An invisible list containing:
  \item{num_tiles}{Number of tiles covered by the shape}
  \item{tile_ids}{Vector of tile IDs covered by the shape}
  \item{available_features}{Vector of available features for the covered tiles}
  \item{area}{Area of the shape in hectares}
  \item{bbox}{Bounding box of the shape (xmin, xmax, ymin, ymax)}
  \item{overlapping_countries}{Vector of countries that overlap with the shape}
  \item{country_groups}{Vector of country groups the shape belongs to}

List containing spatial information (invisibly)
}
\description{
Get Information About a Spatial Shape or Country
}
