% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Get Information About a Spatial Shape or Country}
\usage{
get_info(shape_or_iso, ff_dir = NULL, verbose = TRUE)
}
\arguments{
\item{shape_or_iso}{Either a SpatVector object representing a spatial shape,
or a 3-digit ISO country code as a character string.}

\item{ff_dir}{Character string. The directory path to the ForestForesight data.
If NULL, available features will not be determined. Default is NULL.}

\item{verbose}{Logical. If TRUE, prints the information to the console.
Default is TRUE.}
}
\value{
An invisible list containing:
  \item{num_tiles}{Number of tiles covered by the shape}
  \item{tile_ids}{Vector of tile IDs covered by the shape}
  \item{available_features}{Vector of available features for the covered tiles}
  \item{area}{Area of the shape in hectares}
  \item{bbox}{Bounding box of the shape (xmin, xmax, ymin, ymax)}
  \item{overlapping_countries}{Vector of countries that overlap with the shape}
  \item{country_groups}{Vector of country groups the shape belongs to}
}
\description{
This function provides information about a given spatial shape or country,
including the number of tiles it covers, available features, area, bounding box,
overlapping countries, and country groups.
}
\details{
If a 3-digit ISO code is provided, the function uses the 'countries' dataset
from the ForestForesight package to get the corresponding country shape.
The function then determines which Global Forest Watch (GFW) tiles intersect
with the given shape, calculates the area, finds the bounding box, determines
overlapping countries, and identifies the country groups. If a ForestForesight
directory is provided, it also determines the available features for the
intersecting tiles.
}
\examples{
\dontrun{
# Using an ISO code
result_iso <- get_info("BRA", ff_dir = "/path/to/ForestForesight", verbose = TRUE)

# Using a SpatVector object
my_shape <- vect(your_shapefile_path)
result_shape <- get_info(my_shape, ff_dir = "/path/to/ForestForesight", verbose = TRUE)

# Accessing results
num_tiles <- result_shape$num_tiles
tile_ids <- result_shape$tile_ids
overlapping_countries <- result_shape$overlapping_countries
country_groups <- result_shape$country_groups
}

}
