% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster.R
\name{get_raster}
\alias{get_raster}
\title{Retrieve raster files based on specified criteria.}
\usage{
get_raster(
  datafolder,
  date,
  feature,
  tile = NULL,
  shape = NULL,
  return_raster = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{datafolder}{A character string specifying the path to the data folder where raster files are stored.
This should be in the preprocessed folder of Forest Foresight.}

\item{date}{A Date object representing the date for which raster files are to be retrieved.}

\item{feature}{A character string specifying the feature of interest to filter raster files.}

\item{tile}{A character string specifying the pattern to filter specific tiles of raster files.}

\item{shape}{A spatial shape that can be used instead of tile to return the files for one or more files.}

\item{return_raster}{Logical. Whether it should return the (mosaiced) raster instead of just the file names.}

\item{verbose}{Logical. Whether print statements should be output.}
}
\value{
A character vector containing the file paths of the selected raster files.
}
\description{
This function retrieves raster files from a specified data folder based on the given date, feature, and tile pattern.
}
\examples{
\dontrun{
datafolder <- "/path/to/data"
date <- as.Date("2022-01-01")
feature <- "initialforestcover"
tile <- "00N_070W"
get_raster(datafolder, date, feature, tile)
}
# Returns: Vector of file paths to selected raster files.
}
