% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict_raster.R
\name{train_predict_raster}
\alias{train_predict_raster}
\title{Train and Predict Raster}
\usage{
train_predict_raster(
  shape = NULL,
  country = NULL,
  prediction_date,
  ff_folder,
  train_start = NULL,
  train_end = NULL,
  save_path = NULL,
  trained_model = NULL,
  ff_prep_params = NULL,
  ff_train_params = NULL,
  threshold = 0.5,
  accuracy_csv = NA,
  overwrite = F,
  verbose = T
)
}
\arguments{
\item{shape}{Spatial object representing the shapefile. Either `shape` or `country` should be given.}

\item{country}{ISO3 country code. Either `shape` or `country` should be given.}

\item{prediction_date}{Date for prediction in "YYYY-MM-DD" format.}

\item{ff_folder}{Folder directory containing the input data.}

\item{train_start}{Starting date for training data in "YYYY-MM-DD" format. Default is "2022-07-01".}

\item{train_end}{Ending date for training data in "YYYY-MM-DD" format. Default is "2023-07-01".}

\item{save_path}{The path for saving the model.}

\item{trained_model}{Pre-trained model. If NULL, the function will train a model. Default is NULL.}

\item{ff_prep_params}{list of parameters for data preprocessing.}

\item{ff_train_params}{list of parameters for model training.}

\item{accuracy_csv}{Path to save accuracy metrics in CSV format. Default is NA (no CSV output).}

\item{overwrite}{Logical value indicating whether to overwrite existing files. Default is FALSE.}

\item{verbose}{Logical value indicating whether to display progress messages. Default is TRUE.}
}
\value{
A raster object containing the predicted values.
}
\description{
This function trains a model using training data and then predicts raster values using the trained model.
}
\examples{
# Example usage:
train_predict_raster(prediction_date = "2023-01-01", ff_folder = "path/to/folder")

}
